dont let users delete themselves
Summary:
- prevent users from deleting themselves on the backend
- refresh Manage Users view when a delete or bulk delete fails
- update notify.js to return useful promises
Fixes T906

Test Plan:
Try deleting users individually, including yourself.
Try deleting various combinations of other users and yourself in bulk.

Reviewers: rjmackay, vladimir, shadowhand

Reviewed By: shadowhand

Subscribers: srutto

Maniphest Tasks: T906

Differential Revision: https://phabricator.ushahidi.com/D435

+++ application/classes/Controller/Api/Users.php
@@ -262,14 +262,24 @@ class Controller_Api_Users extends Ushahidi_Api {
 	{
 		$user = $this->resource();
 		$this->_response_payload = array();
-		if ( $user->loaded() )
+
+		if ($user->loaded())
 		{
-			// Return the user we just deleted (provides some confirmation)
-			$this->_response_payload = $this->view->render($user);
-			$this->_response_payload['allowed_methods'] = $this->_allowed_methods();
-			$user->delete();
+			if ($user->id === $this->user->id)
+			{
+				throw new HTTP_Exception_400('Validation Error: \':error\'', array(
+					':error' => 'You may not delete yourself',
+				));
+			}
+			else
+			{
+				// Return the user we just deleted (provides some confirmation)
+				$this->_response_payload = $this->view->render($user);
+				$this->_response_payload['allowed_methods'] = $this->_allowed_methods();
+				$user->delete();
+			}
 		}
-		}
+	}
 
 
 	/**
@@ -326,7 +336,7 @@ class Controller_Api_Users extends Ushahidi_Api {
 		catch(ORM_Validation_Exception $e)
 		{
 			throw new HTTP_Exception_400('Validation Error: \':errors\'', array(
-					':errors' => implode(', ', Arr::flatten($e->errors('models'))),
+				':errors' => implode(', ', Arr::flatten($e->errors('models'))),
 			));
 		}
 	}
